Move Kafka connector to correct folder
