+++ messaging/connectors/kafka/pom.xml
@@ -24,7 +24,7 @@
 
     <parent>
         <groupId>io.helidon.messaging</groupId>
-        <artifactId>helidon-messaging-project</artifactId>
+        <artifactId>helidon-messaging-connectors-project</artifactId>
         <version>3.0.0-SNAPSHOT</version>
     </parent>
