+++ messaging/connectors/kafka/src/main/resources/META-INF/beans.xml
