+++ messaging/connectors/kafka/src/main/resources/META-INF/helidon/native-image/reflection-config.json
