+++ messaging/connectors/pom.xml
@@ -31,6 +31,7 @@
     <packaging>pom</packaging>
 
     <modules>
+        <module>kafka</module>
         <module>jms</module>
         <module>aq</module>
         <module>jms-shim</module>
