+++ messaging/pom.xml
@@ -35,6 +35,5 @@
     <modules>
         <module>messaging</module>
         <module>connectors</module>
-        <module>kafka</module>
     </modules>
 </project>
