+++ tests/integration/kafka/src/test/java/io/helidon/messaging/connectors/kafka/KafkaSubscriberTckTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2020, 2021 Oracle and/or its affiliates.
+ * Copyright (c) 2020, 2022 Oracle and/or its affiliates.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,7 +16,7 @@
 
 package io.helidon.messaging.connectors.kafka;
 
-import java.util.Arrays;
+import java.util.List;
 import java.util.UUID;
 
 import org.apache.kafka.clients.producer.Producer;
@@ -40,10 +40,11 @@ class KafkaSubscriberTckTest extends SubscriberBlackboxVerification<Message<Stri
     }
 
     @Override
+    @SuppressWarnings("unchecked")
     public Subscriber<Message<String>> createSubscriber() {
         Producer<Object, String> producer = Mockito.mock(Producer.class);
         return KafkaSubscriber.<Object, String>builder().producerSupplier(() -> producer)
-                .topics(Arrays.asList("topic")).backpressure(1).build();
+                .topics(List.of("topic")).backpressure(1).build();
     }
 
 }
